refactor: organize component files and
- prompt course and/or assignment selections if they haven't been made
  yet
- #120
