+++ frontend/src/components/AssignmentSelectionMenu.tsx
@@ -1,9 +1,9 @@
 import { ReactElement, useEffect, useState } from "react";
 import { Assignment, PaletteAPIResponse } from "palette-types";
 import { useFetch } from "@hooks";
-import { useCourse } from "../../context";
-import { useAssignment } from "../../context/AssignmentProvider.tsx";
-import LoadingDots from "../../components/LoadingDots.tsx";
+import { useCourse } from "../context";
+import { useAssignment } from "../context/AssignmentProvider.tsx";
+import LoadingDots from "./LoadingDots.tsx";
 
 export default function AssignmentSelectionMenu({
   onSelect,
